Migrated load account data from local storage to jest
